+++ apps/text_file.py
@@ -2,7 +2,6 @@ import sys
 import os
 
 sys.path.append("..\src")
-
 from sentimeter.sentimeter import Sentimeter
 from sentimeter.sources import TextSource
 from sentimeter.backends import AIRemoteBackend
